+++ src/main/java/com/bitmart/api/request/account/pub/AccountCurrenciesRequest.java
@@ -1,5 +1,6 @@
 package com.bitmart.api.request.account.pub;
 
+import com.bitmart.api.annotations.ParamKey;
 import com.bitmart.api.request.Auth;
 import com.bitmart.api.request.CloudRequest;
 import com.bitmart.api.request.Method;
@@ -10,18 +11,18 @@ import lombok.experimental.Accessors;
 
 /**
  * Get Currencies
- * <p>
- * GET <a href="https://api-cloud.bitmart.com/account/v1/currencies">
- *     https://api-cloud.bitmart.com/account/v1/currencies</a>
- * <br>
- * @see <a href="https://developer-pro.bitmart.com/en/spot/#get-currencies">
- *     BitMart Document</a>
  */
 @EqualsAndHashCode(callSuper = true)
 @Data
 @ToString
 @Accessors(chain = true)
 public class AccountCurrenciesRequest extends CloudRequest {
+
+    /**
+     *  currency, e.g., 'BTC'
+     */
+    @ParamKey("currency")
+    private String currency;
 
     public AccountCurrenciesRequest() {
         super("/account/v1/currencies", Method.GET, Auth.NONE);