+++ src/test/java/com/bitmart/examples/spot/trade/NewOrder.java
@@ -7,6 +7,7 @@ import com.bitmart.api.common.CloudException;
 import com.bitmart.api.common.CloudResponse;
 import com.bitmart.api.key.CloudKey;
 import com.bitmart.api.request.spot.prv.SubmitOrderRequest;
+import com.bitmart.examples.Example;
 import lombok.extern.slf4j.Slf4j;
 
 import static com.bitmart.api.common.GlobalConst.CLOUD_URL;
@@ -14,9 +15,9 @@ import static com.bitmart.api.common.GlobalConst.CLOUD_URL;
 @Slf4j
 public class NewOrder {
 
-    private static final String API_KEY = "your_api_key";
-    private static final String API_SECRET = "your_secret_key";
-    private static final String API_MEMO = "your_memo";
+    private static final String API_KEY = Example.YOUR_API_KEY;
+    private static final String API_SECRET = Example.YOUR_API_SECRET;
+    private static final String API_MEMO = Example.YOUR_API_MEMO;
 
     public static void main(String[] args) {
         Call call = new Call(new CloudContext(CLOUD_URL, new CloudKey(API_KEY, API_SECRET, API_MEMO)));
@@ -28,7 +29,8 @@ public class NewOrder {
                     .setSymbol("BTC_USDT")
                     .setClient_order_id("def123123123")
                     .setPrice("800")
-                    .setSize("0.1"));
+                    .setSize("0.1")
+                    .setStpMode("cancel_taker")); // cancel_taker
             System.out.println(cloudResponse);
 
             // Special Parameters for Market Buy Orders (type=market, side=buy)